% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbonInput.R
\name{calcCarbonInput}
\alias{calcCarbonInput}
\title{calcCarbonInput}
\usage{
calcCarbonInput(landtype = "crop")
}
\arguments{
\item{landtype}{'crop' for cropland, 'natveg' for all the rest}
}
\value{
magpie object in cellular resolution
}
\description{
This function compiles carbon inputs as well as lignin and nitrogen content for mineral soils
using the steady-state method (Tier 2) of the 2019 Refinement to the 2006 IPP Guidelines
for National Greenhouse Gas Inventories
}
\examples{
\dontrun{ calcOutput("CarbonInput", aggregate = FALSE) }

}
\author{
Kristine Karstens
}
