% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbonManure.R
\name{calcCarbonManure}
\alias{calcCarbonManure}
\title{calcCarbonManure}
\usage{
calcCarbonManure()
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Calculates carbon input from manure for cropland.
}
\examples{

\dontrun{
calcOutput("CarbonManure")
}
}
\author{
Kristine Karstens
}
