% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbonResidues.R
\name{calcCarbonResidues}
\alias{calcCarbonResidues}
\title{calcCarbonResidues}
\usage{
calcCarbonResidues(
  yieldscenario = "default",
  rec.scenario = "default",
  res.scenario = "default"
)
}
\arguments{
\item{yieldscenario}{yield scenario}

\item{rec.scenario}{recycling scenario}

\item{res.scenario}{residue scenario}
}
\value{
List of magpie object with results on cellular level, weight on cellular level, unit and description.
}
\description{
Calculates carbon input from residues for cropland soils.
}
\examples{

\dontrun{
calcOutput("CarbonResidues")
}
}
\author{
Kristine Karstens
}
