% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDecay.R
\name{calcDecay}
\alias{calcDecay}
\title{calcDecay}
\usage{
calcDecay()
}
\value{
magpie object in cellular resolution
}
\description{
This function wraps together the decay rate for allSOC sub-pool per year for mineral soils
using the steady-state method (Tier 2) of the 2019 Refinement to the 2006 IPP Guidelines
for National Greenhouse Gas Inventories
}
\examples{
\dontrun{ calcOutput("Decay", aggregate = FALSE) }

}
\author{
Kristine Karstens
}
