% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigationMonth.R
\name{calcIrrigationMonth}
\alias{calcIrrigationMonth}
\title{calcIrrigationMonth}
\usage{
calcIrrigationMonth()
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the length of irrigation period crop type specific
}
\examples{
\dontrun{ calcOutput("IrrigationMonth", aggregate = FALSE) }

}
\author{
Kristine Karstens
}
