% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLanduse.R
\name{calcLanduse}
\alias{calcLanduse}
\title{calcLanduse}
\usage{
calcLanduse(period = "states_1900to2010", output = "total")
}
\arguments{
\item{period}{select historical period (handed over to readLUH2v2)}

\item{output}{"total" or "change"}
}
\value{
List of magpie object with results on cellular level, weight, unit and description.
}
\description{
Calculates the cellular MAgPIE landuse area based on LUH2v2
}
\examples{
\dontrun{
calcOutput("Landuse", aggregate = FALSE)
}

}
\author{
Kristine Karstens
}
