% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLanduseChange.R
\name{calcLanduseChange}
\alias{calcLanduseChange}
\title{calcLanduseChange}
\usage{
calcLanduseChange(cfg = NULL)
}
\arguments{
\item{cfg}{general configuration}
}
\value{
List of magpie object with results on country or cellular level, weight, unit and description.
}
\description{
Calculates the cellular MAgPIE landuse change area based on LUH2v2 or LanduseInitialisation data.
}
\examples{

\dontrun{
calcOutput("LanduseChange")
}
}
\author{
Kristine Karstens
}
