% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLitterSoilinput.R
\name{calcLitterSoilinput}
\alias{calcLitterSoilinput}
\title{calcLitterSoilinput}
\usage{
calcLitterSoilinput(
  lpjmlNatveg = "LPJmL4_for_MAgPIE_44ac93de",
  climatetype = "GSWP3-W5E5:historical",
  fixFpc = FALSE
)
}
\arguments{
\item{lpjmlNatveg}{Switch between LPJmL natveg versionstop}

\item{climatetype}{Switch between different climate scenarios}

\item{fixFpc}{if TRUE using fixed (old) fpc data}
}
\value{
List of magpie object with results on cellular level, weight on cellular level, unit and description.
}
\description{
Calculates Carbon Input from litter to the soil pools of the three pool model (Tier 2 IPCC)
}
\examples{
\dontrun{
calcOutput("calcLitterSoilinput", aggregate = FALSE)
}
}
\author{
Kristine Karstens
}
