% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcParamManure.R
\name{calcParamManure}
\alias{calcParamManure}
\title{calcParamManure}
\usage{
calcParamManure(source = "IPCC")
}
\arguments{
\item{source}{"IPCC"          for IPCC Guideline values}
}
\value{
List of magpie object with results on global level, unit and description.
}
\description{
Bring all parameter settings (lignin, nitrogen) for residues together
}
\examples{
\dontrun{
calcOutput("ParamManure")
}
}
\author{
Kristine Karstens
}
