% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcParamResidues.R
\name{calcParamResidues}
\alias{calcParamResidues}
\title{calcParamResidues}
\usage{
calcParamResidues(source = "IPCC+woody")
}
\arguments{
\item{source}{"IPCC"          for IPCC Guideline values
"IPCC+woody"    for IPCC Guideline values + Feedipedia for woody}
}
\value{
List of magpie object with results on global level, unit and description.
}
\description{
Bring all parameter settings (lignin, nitrogen) for residues together
}
\examples{
\dontrun{
calcOutput("ParamResidues")
}
}
\author{
Kristine Karstens
}
