% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPassiveDecay.R
\name{calcPassiveDecay}
\alias{calcPassiveDecay}
\title{calcPassiveDecay}
\usage{
calcPassiveDecay(landtype = "crop")
}
\arguments{
\item{landtype}{'crop' for cropland, 'natveg' for all the rest}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the decay rate for passive SOC sub-pool per year for mineral soils
using the steady-state method (Tier 2) of the 2019 Refinement to the 2006 IPP Guidelines
for National Greenhouse Gas Inventories
}
\examples{
\dontrun{ calcOutput("PassiveDecay", aggregate = FALSE) }

}
\author{
Kristine Karstens
}
