% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPassiveSteadyState.R
\name{calcPassiveSteadyState}
\alias{calcPassiveSteadyState}
\title{calcPassiveSteadyState}
\usage{
calcPassiveSteadyState(landtype = "crop")
}
\arguments{
\item{landtype}{'crop' for cropland, 'natveg' for all the rest}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the steady state passive sub-pool SOC stock for mineral soils
using the steady-state method (Tier 2) of the 2019 Refinement to the 2006 IPP Guidelines
for National Greenhouse Gas Inventories
}
\examples{
\dontrun{ calcOutput("PassiveSteadyState", aggregate = FALSE) }

}
\author{
Kristine Karstens
}
