% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSandFrac.R
\name{calcSandFrac}
\alias{calcSandFrac}
\title{calcSandFrac}
\usage{
calcSandFrac()
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the fraction of 0-30cm soil mass that is sand (0.050 - 2mm particles)
}
\examples{
\dontrun{ calcOutput("SandFrac", aggregate = FALSE) }

}
\seealso{
\code{\link[mrcommons]{readSoilGrids}}
}
\author{
Kristine Karstens
}
