% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSOCDebt.R
\name{calcSOCDebt}
\alias{calcSOCDebt}
\title{calcSOCDebt}
\usage{
calcSOCDebt()
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates and return SOC stocks and debts from Soil Carbon Debt github repo
}
\examples{
 \dontrun{ calcOutput("calcSOCDebt", aggregate = FALSE) }
}
\seealso{
\code{\link[mrvalidation]{readSoilCarbonDebt}}
}
\author{
Kristine Karstens
}
