% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSOCPointData.R
\name{calcSOCPointData}
\alias{calcSOCPointData}
\title{calcSOCPointData}
\usage{
calcSOCPointData()
}
\value{
magpie object in cellular resolution
}
\description{
This function return SOC point data for validation from Sanderman et al., 2017
}
\examples{
 \dontrun{ calcOutput("calcSOCPointData", aggregate = FALSE) }
}
\seealso{
\code{\link[mrsoil]{readSanderman}}
}
\author{
Kristine Karstens
}
