% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSoilCarbonSpinup.R
\name{calcSoilCarbonSpinup}
\alias{calcSoilCarbonSpinup}
\title{calcSoilCarbonSpinup}
\usage{
calcSoilCarbonSpinup(spinup_start = 1510, cfg_default = NULL)
}
\arguments{
\item{spinup_start}{start year of spinup (model_start-1-spinup_start has to be a multiple of recycling length)}

\item{cfg_default}{run configuration for default setup (used for spinup setup)}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates SOC states
}
\examples{
\dontrun{
calcOutput("SoilCarbon")
}
}
\author{
Kristine Karstens
}
