% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSteadyState.R
\name{calcSteadyState}
\alias{calcSteadyState}
\title{calcSteadyState}
\usage{
calcSteadyState(
  lpjmlNatveg = "LPJmL4_for_MAgPIE_44ac93de",
  climatetype = "GSWP3-W5E5:historical"
)
}
\arguments{
\item{lpjmlNatveg}{Switch between LPJmL natveg versionstop}

\item{climatetype}{Switch between different climate scenarios}
}
\value{
magpie object in cellular resolution
}
\description{
This function wraps together the steady state for all sub-pool SOC stock for mineral soils
             using the steady-state method (Tier 2) of the 2019 Refinement to the 2006 IPP Guidelines
             for National Greenhouse Gas Inventories
}
\examples{
\dontrun{
calcOutput("SteadyState", aggregate = FALSE)
}

}
\author{
Kristine Karstens
}
