% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTillageArea.R
\name{calcTillageArea}
\alias{calcTillageArea}
\title{calcTillageArea}
\usage{
calcTillageArea(tillage = "historicNoTill")
}
\arguments{
\item{tillage}{'historicNoTill' (default) or 'ruleBased'; histroricNoTill' is based
on the same assunmptions as 'ruleBased', but additionally accounts for
historic data on no tillage areas based on Porwolliks datasets.
'ruleBased' assumes full tillage for annual crops, and reduced tillage for perennials.}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the tillage area shares for different tillage types
}
\examples{
\dontrun{
calcOutput("TillageArea", aggregate = FALSE)
}

}
\author{
Kristine Karstens
}
