% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTransferActive2Slow.R
\name{calcTransferActive2Slow}
\alias{calcTransferActive2Slow}
\title{calcTransferActive2Slow}
\usage{
calcTransferActive2Slow()
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the fraction of active SOC sub-pool decay products
transferred to the slow SOC sub-pool for mineral soils using the steady-state method (Tier 2)
of the 2019 Refinement to the 2006 IPP Guidelines for National Greenhouse Gas Inventories
}
\examples{
\dontrun{ calcOutput("TransferActive2Slow", aggregate = FALSE) }

}
\author{
Kristine Karstens
}
