% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWaterEffectDecomposition.R
\name{calcWaterEffectDecomposition}
\alias{calcWaterEffectDecomposition}
\title{calcWaterEffectDecomposition}
\usage{
calcWaterEffectDecomposition(irrigation = "mixed")
}
\arguments{
\item{irrigation}{irrigation type to de considered. Default (mixed) is historic irrigation area shares
to calculate area weighted mean over rainfed and irrigated factors. Other options: rainfed, irrigated}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the water effect on decomposition for mineral soils
using the steady-state method (Tier 2) of the 2019 Refinement to the 2006 IPP Guidelines
for National Greenhouse Gas Inventories for rainfed and irrigated systems
}
\examples{
\dontrun{ calcOutput("WaterEffectDecomposition", aggregate = FALSE) }

}
\author{
Kristine Karstens
}
