% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctPorwolliksGriddedTillage.R
\name{correctPorwolliksGriddedTillage}
\alias{correctPorwolliksGriddedTillage}
\title{correctPorwolliksGriddedTillage}
\usage{
correctPorwolliksGriddedTillage(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read historical tillage data set based on (https://www.earth-syst-sci-data.net/11/823/2019/)
}
\examples{

\dontrun{
  readSource("PorwolliksGriddedTillage", convert="onlycorrect")
}

}
\seealso{
\code{\link[mrcommons]{readLPJmL}},
\code{\link[lpjclass]{read.LPJ_input}}
}
\author{
Kristine Karstens
}
