% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadCENTURY.R
\name{downloadCENTURY}
\alias{downloadCENTURY}
\title{downloadCENTURY}
\usage{
downloadCENTURY()
}
\description{
This function download the CENTURY model (available at 'https://www2.nrel.colostate.edu/projects/century/century.tar.Z')
}
\examples{
\dontrun{ downloadSource("CENTURY") }
}
\seealso{
\code{\link{downloadCENTURY}}
}
\author{
Kristine Karstens
}
