% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadIPCCSoil.R
\name{downloadIPCCSoil}
\alias{downloadIPCCSoil}
\title{downloadIPCCSoil}
\usage{
downloadIPCCSoil()
}
\value{
magpie object with global parameters
}
\description{
This function download the IPCC table data
}
\examples{
\dontrun{ downloadSource("IPCCSoil") }

}
\seealso{
\code{\link{readIPCCSoil}}
}
\author{
Kristine Karstens
}
