% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrsoil-package.R
\docType{package}
\name{mrsoil-package}
\alias{mrsoil}
\alias{mrsoil-package}
\title{mrsoil: MadRat Soil Organic Carbon Budget Library}
\description{
This packages provides functions for soil organic carbon budget for mineral soils using the steady-state method (Tier 2) of the 2019 Refinement to the 2006 IPP Guidelines for National Greenhouse Gas Inventories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrsoil}
  \item \doi{10.5281/zenodo.4317933}
  \item Report bugs at \url{https://github.com/pik-piam/mrsoil/issues}
}

}
\author{
\strong{Maintainer}: Kristine Karstens \email{karstens@pik-potsdam.de}

Authors:
\itemize{
  \item Jan Philipp Dietrich \email{dietrich@pik-potsdam.de}
}

}
