% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBrovkin.R
\name{readBrovkin}
\alias{readBrovkin}
\title{readBrovkin}
\usage{
readBrovkin()
}
\value{
List of magpie objects with results on global level
}
\description{
Read parameter files from Brovkin Paper (doi:10.5194/bg-9-565-2012)
}
\examples{
\dontrun{
readSource("Brovkin", convert = FALSE)
}

}
\author{
Kristine Karstens
}
