% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCENTURY.R
\name{readCENTURY}
\alias{readCENTURY}
\title{readCENTURY}
\usage{
readCENTURY(subtype = "tree")
}
\arguments{
\item{subtype}{Switch between different parameter files}
}
\value{
List of magpie objects with results on global level
}
\description{
Read parameter files in CENTURY model and return n/c- and lg/c-ratios
}
\examples{
\dontrun{
  readSource("CENTURY", subtype="tree")
}

}
\author{
Kristine Karstens
}
