% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIPCCSoil.R
\name{readIPCCSoil}
\alias{readIPCCSoil}
\title{readIPCCSoil}
\usage{
readIPCCSoil(subtype = "steady_state")
}
\arguments{
\item{subtype}{Switch between different input}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read IPCC Guideline tables
}
\examples{

\dontrun{
readSource("IPCCSoil")
}

}
\seealso{
\code{\link{downloadIPCCSoil}},
}
\author{
Kristine Karstens
}
