% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLPJmL_par.R
\name{readLPJmL_par}
\alias{readLPJmL_par}
\title{readLPJmL_par}
\usage{
readLPJmL_par(subtype = "pft_lpjml4")
}
\arguments{
\item{subtype}{Switch between different parameter files}
}
\value{
List of magpie objects with results on global level
}
\description{
Read parameter files from LPJmL model
}
\examples{
\dontrun{
readSource("LPJmL_par", subtype = "pft_lpjml4")
}

}
\author{
Kristine Karstens
}
