% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSanderman.R
\name{readSanderman}
\alias{readSanderman}
\title{readSanderman}
\usage{
readSanderman()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read validation data from Sanderman et al., 2017
}
\examples{
\dontrun{
readSource("Sanderman")
}

}
\author{
Kristine Karstens
}
