% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSOCbudgetParam.R
\name{readSOCbudgetParam}
\alias{readSOCbudgetParam}
\title{readSOCbudgetParam}
\usage{
readSOCbudgetParam(subtype = "default")
}
\arguments{
\item{subtype}{Switch between different input}
}
\description{
Read SOC budget parameters for sensitivity analysis
}
\examples{

\dontrun{
readSource("readSOCbudgetParam")
}

}
\author{
Kristine Karstens
}
