% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTillage.R
\name{readTillage}
\alias{readTillage}
\title{readTillage}
\usage{
readTillage()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read historical tillage data set based on (https://www.earth-syst-sci-data.net/11/823/2019/)
}
\examples{

\dontrun{
  readSource("Tillage")
}

}
\seealso{
\code{\link[mrcommons]{readLPJmL}},
\code{\link[lpjclass]{read.LPJ_input}}
}
\author{
Kristine Karstens
}
