% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolBindSubtypes.R
\name{toolBindSubtypes}
\alias{toolBindSubtypes}
\title{toolBindSubtypes}
\usage{
toolBindSubtypes(subtype_name, subtype_list, type, ...)
}
\arguments{
\item{subtype_name}{subtype name that should be binded}

\item{subtype_list}{subtype list that should be binded}

\item{type}{calcOutput-type}

\item{...}{addition arguments handed over to calcOutput call}
}
\value{
magpie object with global parameters
}
\description{
This function mbinds subtypes for different functions
}
\examples{
\dontrun{ downloadSource("IPCCSoil") }

}
\seealso{
\code{\link{readIPCCSoil}}
}
\author{
Kristine Karstens
}
