% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCarbonInputMultiplier.R
\name{toolCarbonInputMultiplier}
\alias{toolCarbonInputMultiplier}
\title{toolCarbonInputMulitplier}
\usage{
toolCarbonInputMultiplier(inputProp, soilParam, f4act2slo, f2struc2act)
}
\arguments{
\item{inputProp}{input properties: lignin to c (LC) and nitrogen to c (NC) ratio
on cellular or global resolution, with or without time resolution}

\item{soilParam}{soil model parameters (via readSource("IPCCSoil"))}

\item{f4act2slo}{stabilization efficiencies for active pool decay products entering the passive pool
(this factor is sand content depending and therefore spatial explicit)}

\item{f2struc2act}{stabilization efficiencies for structural decay products entering the active pool
(this factor is tillage type depending and can but not must be spatial explicit)}
}
\value{
magpie object in input resolution
}
\description{
This function compiles carbon inputs multipliers for steady state calculations for mineral soils
             using the steady-state method (Tier 2) of the 2019 Refinement to the 2006 IPP Guidelines
             for National Greenhouse Gas Inventories
}
\author{
Kristine Karstens
}
