% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolFreezeAverage.R
\name{toolFreezeAverage}
\alias{toolFreezeAverage}
\title{toolFreezeAverage}
\usage{
toolFreezeAverage(x, freeze_year)
}
\arguments{
\item{x}{variable to freeze}

\item{freeze_year}{start of freeze period}
}
\value{
magpie object wth froozen effect
}
\description{
Calculates the 10 year average and freezes the effect for all years after
}
\author{
Kristine Karstens
}
