% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolSoilCarbonCycling.R
\name{toolSoilCarbonCycling}
\alias{toolSoilCarbonCycling}
\title{toolSoilCarbonCycling}
\usage{
toolSoilCarbonCycling(
  SoilCarbon,
  SoilCarbonSteadyState,
  Decay,
  Landuse,
  LanduseChange
)
}
\arguments{
\item{SoilCarbon}{soil carbon to be filled}

\item{SoilCarbonSteadyState}{steadystates}

\item{Decay}{decay rates}

\item{Landuse}{landuse}

\item{LanduseChange}{landuse change}
}
\value{
magpie object with global parameters
}
\description{
This function cycles the carbon on an annual basis between the different soil pools
}
\author{
Kristine Karstens
}
