% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbonLitter.R
\name{calcCarbonLitter}
\alias{calcCarbonLitter}
\title{calcCarbonLitter}
\usage{
calcCarbonLitter(
  lpjml = "LPJmL4_for_MAgPIE_44ac93de",
  climatetype = "GSWP3-W5E5:historical",
  mode = "historicalSpinup"
)
}
\arguments{
\item{lpjml}{Switch between LPJmL natveg versionstop}

\item{climatetype}{Switch between different climate scenarios}

\item{mode}{"historicalSpinup" for historical period and
"magpieInput" for future
"magpieInput+fixedFPC" for future but given data restriction on fpc data}
}
\value{
List of magpie object with results on cellular level, weight on cellular level, unit and description.
}
\description{
Calculates Carbon Input from litter
}
\examples{
\dontrun{
calcOutput("CarbonLitter", aggregate = FALSE)
}
}
\author{
Kristine Karstens
}
