% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbonLitter.R
\name{calcCarbonLitter}
\alias{calcCarbonLitter}
\title{calcCarbonLitter}
\usage{
calcCarbonLitter(litter_param = "default", climate_scen = "default")
}
\arguments{
\item{litter_param}{litter scenario}

\item{climate_scen}{climate configuration}
}
\value{
List of magpie object with results on cellular level, weight on cellular level, unit and description.
}
\description{
Calculates Carbon Input from litter
}
\examples{

\dontrun{
  calcOutput("CarbonLitter")
}
}
\author{
Kristine Karstens
}
