% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbonResidues.R
\name{calcCarbonResidues}
\alias{calcCarbonResidues}
\title{calcCarbonResidues}
\usage{
calcCarbonResidues(scenario = "default")
}
\arguments{
\item{scenario}{define scenario switch for sensititvy analysis.
'default' for historic assumptions
'freezeXXXX' for frozen residue recycling rates from year XXXX
'yieldsXXXX' for frozen yields for res calculations from year XXXX}
}
\value{
List of magpie object with results on cellular level, weight on cellular level, unit and description.
}
\description{
Calculates carbon input from residues for cropland soils.
}
\examples{

\dontrun{
calcOutput("CarbonResidues")
}
}
\author{
Kristine Karstens
}
