% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbonResidues.R
\name{calcCarbonResidues}
\alias{calcCarbonResidues}
\title{calcCarbonResidues}
\usage{
calcCarbonResidues()
}
\value{
List of magpie object with results on cellular level, weight on cellular level, unit and description.
}
\description{
Calculates carbon input from residues for cropland soils.
}
\examples{
\dontrun{
calcOutput("CarbonResidues")
}
}
\author{
Kristine Karstens
}
