% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSoilCarbon.R
\name{calcSoilCarbon}
\alias{calcSoilCarbon}
\title{calcSoilCarbon}
\usage{
calcSoilCarbon(
  output = "actualstate",
  lpjmlNatveg = "LPJmL4_for_MAgPIE_44ac93de",
  climatetype = "GSWP3-W5E5:historical"
)
}
\arguments{
\item{output}{"actualstate" (default), "carbontransfer", "interstate", "naturalstate"}

\item{lpjmlNatveg}{Switch between LPJmL natveg versionstop}

\item{climatetype}{Switch between different climate scenarios}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates SOC states
}
\examples{
\dontrun{
calcOutput("SoilCarbon")
}
}
\author{
Kristine Karstens
}
