% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSoilCarbon.R
\name{calcSoilCarbon}
\alias{calcSoilCarbon}
\title{calcSoilCarbon}
\usage{
calcSoilCarbon(
  spinup_start = "1510",
  output = "full",
  cfg = NULL,
  cfg_default = NULL
)
}
\arguments{
\item{spinup_start}{start of initialization with natural vegetation stocks}

\item{output}{'full' - all important soil related values, 'reduced' - just SOC state values}

\item{cfg}{run configuration}

\item{cfg_default}{run configuration for default setup (used for spinup setup)}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates SOC states
}
\examples{
\dontrun{
calcOutput("SoilCarbon")
}
}
\author{
Kristine Karstens
}
