% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSoilCarbonSpinup.R
\name{calcSoilCarbonSpinup}
\alias{calcSoilCarbonSpinup}
\title{calcSoilCarbonSpinup}
\usage{
calcSoilCarbonSpinup(model_start = 1901, cfg_default = NULL)
}
\arguments{
\item{model_start}{start year}

\item{cfg_default}{run configuration for default setup (used for spinup setup)}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates SOC states
}
\examples{
\dontrun{
calcOutput("SoilCarbon")
}
}
\author{
Kristine Karstens
}
