% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTempEffectDecomposition.R
\name{calcTempEffectDecomposition}
\alias{calcTempEffectDecomposition}
\title{calcTempEffectDecomposition}
\usage{
calcTempEffectDecomposition(climate_scen = "default")
}
\arguments{
\item{climate_scen}{climate configuration}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the temperature effect on decomposition for mineral soils
using the steady-state method (Tier 2) of the 2019 Refinement to the 2006 IPP Guidelines
for National Greenhouse Gas Inventories
}
\examples{
\dontrun{ calcOutput("TempEffectDecomposition", aggregate = FALSE) }

}
\author{
Kristine Karstens
}
