% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTillageArea.R
\name{calcTillageArea}
\alias{calcTillageArea}
\title{calcTillageArea}
\usage{
calcTillageArea(tillage = "default")
}
\arguments{
\item{tillage}{'histtill' or 'mixedtill';
'mixedtill' assumes full tillage for annual crops, and reduced tillage for perennials.
'histtill' is based on the same assunmptions as 'mixedtill', but additionally accounts for
historic data on no tillage areas based on Porwolliks datasets for annual no tillage areas
(other option 'default' for using the current default)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the tillage area shares for different tillage types
}
\examples{
\dontrun{
calcOutput("TillageArea", aggregate = FALSE)
}

}
\author{
Kristine Karstens
}
