% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTillageEffectDecomposition.R
\name{calcTillageEffectDecomposition}
\alias{calcTillageEffectDecomposition}
\title{calcTillageEffectDecomposition}
\usage{
calcTillageEffectDecomposition(tillage = "default")
}
\arguments{
\item{tillage}{tillage type to de considered.
'histtill' is historic tillage area shares based on no
tillage areas from Porwollik together with rule based assumption;
'mixedtill' includes pure rule based assumptions.
Other options: 'fulltill', 'notill', 'reducedtill', 'default'}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the tillage effect on decomposition
             for mineral soils using the steady-state method (Tier 2) of the 2019
             Refinement to the 2006 IPP Guidelines for National Greenhouse Gas Inventories
}
\examples{
\dontrun{
  calcOutput("TillageEffectDecomposition", aggregate = FALSE)
}
}
\author{
Kristine Karstens
}
