% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullCARBONBUDGET.R
\name{fullCARBONBUDGET}
\alias{fullCARBONBUDGET}
\title{fullCARBONBUDGET}
\usage{
fullCARBONBUDGET(rev = 0.1, dev = "", start_year = 1901)
}
\arguments{
\item{rev}{data revision which should be used as input (positive numeric).}

\item{dev}{flag to define carbon budget calculations}

\item{start_year}{start year}
}
\description{
Function that produces the complete cellular data set of the SOC budget
}
\examples{
\dontrun{
  retrieveData("CARBONBUDGET", rev=0, mainfolder="pathtowhereallfilesarestored")
}
}
\seealso{
\code{\link{readSource}},\code{\link{getCalculations}},\code{\link{calcOutput}},\code{\link{setConfig}}
}
\author{
Kristine Karstens
}
