% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCRU.R
\name{readCRU}
\alias{readCRU}
\title{readCRU}
\usage{
readCRU(subtype = "precipitation")
}
\arguments{
\item{subtype}{Switch between diffrent input}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read CRU content
}
\examples{

\dontrun{
  readSource("CRU", subtype="precipitation")
}

}
\seealso{
\code{\link[mrcommons]{readLPJmL}},
\code{\link[lpjclass]{read.LPJ_input}}
}
\author{
Kristine Karstens
}
