% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolLanduseChange.R
\name{toolLanduseChange}
\alias{toolLanduseChange}
\title{toolLanduseChange}
\usage{
toolLanduseChange(landuse, landuse_scen)
}
\arguments{
\item{landuse}{landuse data}

\item{landuse_scen}{landuse configuration}
}
\value{
List of magpie object with results on cellular level
}
\description{
Calculates the cellular MAgPIE landuse change area based on given landuse data and scenario
}
\author{
Kristine Karstens
}
