% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolSoilCarbonCycling.R
\name{toolSoilCarbonCycling}
\alias{toolSoilCarbonCycling}
\title{toolSoilCarbonCycling}
\usage{
toolSoilCarbonCycling(SoilCarbonInit, SoilCarbonSteadyState, Decay, Landuse)
}
\arguments{
\item{SoilCarbonInit}{soil carbon initialization}

\item{SoilCarbonSteadyState}{steadystates}

\item{Decay}{decay rates}

\item{Landuse}{landuse}
}
\value{
magpie object with global parameters
}
\description{
This function cycles the carbon on an annual basis between the different soil pools
}
\author{
Kristine Karstens
}
