% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolSoilCarbonCycling.R
\name{toolSoilCarbonCycling}
\alias{toolSoilCarbonCycling}
\title{toolSoilCarbonCycling}
\usage{
toolSoilCarbonCycling(soilCarbonInit, soilCarbonSteadyState, decay, landuse)
}
\arguments{
\item{soilCarbonInit}{soil carbon initialization}

\item{soilCarbonSteadyState}{steadystates}

\item{decay}{decay rates}

\item{landuse}{landuse}
}
\value{
magpie object with global parameters
}
\description{
This function cycles the carbon on an annual basis between the different soil pools
}
\author{
Kristine Karstens
}
